feat(profile): added a static profile component and placed in explorer
