+++ packages/profiles/src/demo.js
@@ -1,5 +1,6 @@
 import TestComponent from './TestComponent'
 import ProfileChecker from './ProfileChecker'
+import UserProfile from './UserProfile'
 import Profile from './Profile'
 import ProfileForm from './ProfileForm'
 import ProfilePicture from './ProfilePicture'
@@ -7,5 +8,5 @@ import ReactDOM from 'react-dom'
 import React from 'react'
 import AlertDialog from './AlertDialog'
 
-const element = <Profile />
+const element = <Profile address="0x075A8e26b12461acf8B60A32c6943A27d3ed09D0" />
 ReactDOM.render(element, document.querySelector('#react-div'))
