+++ packages/profiles/src/index.js
@@ -1,1 +1,2 @@
+export { default as UserProfile } from './UserProfile'
 export { default as Profile } from './Profile'
